Remove direct dependency on cement prior to install phase in setup.py
The dependency on cement was for get_version functionality.  This logic in cement does not rely on any cement specific logic and was originally copied from Django.  That code has been replicated here to prevent the import of cement prior to its installation.  The Django license text has been included in this file in accordance with the license.